genprotimg: introduce new tool for the creation of PV images
Protected VMs (PVM) are KVM VMs, where KVM can't access the VM's state
like guest memory and guest registers anymore. Instead the PVMs are
mostly managed by a new entity called Ultravisor (UV), which provides
an API, so KVM and the PV can request management actions.

PVMs are encrypted at rest and protected from hypervisor access while
running. They switch from a normal operation into protected mode, so
we can still use the standard boot process to load an encrypted image
and then move it into protected mode.

This commit adds the tooling with the name 'genprotimg' for the
creation of this encrypted image. genprotimg is essentially a program
that takes a kernel, key files, and optionally an initial ram
filesystem, and optionally a file containing the kernel command line
parameters and generates this single loadable image file. This created
image file consists of a concatenation of a plain text boot loader,
the encrypted components for kernel, initrd, and cmdline, and the
integrity-protected PV header, containing metadata necessary for
running the guest in PV mode.

It's possible to use this image file as a kernel for zipl or for a
direct kernel boot using QEMU.

Signed-off-by: Marc Hartmayer <mhartmay@linux.ibm.com>